Convert amounts to NUMERIC in dbt_total_supply.sql
that fixed the rounding

+++ dbt/models/spl-token/dbt_total_supply.sql
@@ -1,3 +1,3 @@
 {{ config(materialized='table') }}
 
-select (SELECT SUM(amount) FROM spl.mints) - (SELECT COALESCE(SUM(amount), 0)  FROM spl.burns) as total_supply
+select (SELECT SUM(amount::NUMERIC) FROM spl.mints) - (SELECT COALESCE(SUM(amount::NUMERIC), 0)  FROM spl.burns) as total_supply
